Moved getGeolocation and Loader into App.js
